refaktoring m_arrCrsc - treba si zavolat a prejst si to

+++ PFD/Infrastructure/CMemberLoadGenerator.cs
@@ -116,14 +116,14 @@ namespace PFD
             m_arrLoadCases = model.m_arrLoadCases;
             m_arrMembers = model.m_arrMembers;
 
-            GirtCrSc = model.m_arrCrSc[(int)EMemberGroupNames.eGirtWall];
-            PurlinCrSc = model.m_arrCrSc[(int)EMemberGroupNames.ePurlin];
+            GirtCrSc = model.m_arrCrSc[EMemberGroupNames.eGirtWall];
+            PurlinCrSc = model.m_arrCrSc[EMemberGroupNames.ePurlin];
             fDistanceGirts = model.fDist_Girt;
             fDistancePurlins = model.fDist_Purlin;
-            ColumnCrSc = model.m_arrCrSc[(int)EMemberGroupNames.eMainColumn];
-            RafterCrSc = model.m_arrCrSc[(int)EMemberGroupNames.eRafter];
-            ColumnCrSc_EF = model.m_arrCrSc[(int)EMemberGroupNames.eMainColumn_EF];
-            RafterCrSc_EF = model.m_arrCrSc[(int)EMemberGroupNames.eRafter_EF];
+            ColumnCrSc = model.m_arrCrSc[EMemberGroupNames.eMainColumn];
+            RafterCrSc = model.m_arrCrSc[EMemberGroupNames.eRafter];
+            ColumnCrSc_EF = model.m_arrCrSc[EMemberGroupNames.eMainColumn_EF];
+            RafterCrSc_EF = model.m_arrCrSc[EMemberGroupNames.eRafter_EF];
 
             fValueLoadColumnDead = -generalLoad.fDeadLoadTotal_Wall;
             fValueLoadRafterDead = -generalLoad.fDeadLoadTotal_Roof;
